+++ artist/field/__init__.py
@@ -1,8 +1,8 @@
 """Bundle all classes that represent physical objects in ARTIST."""
 
-from artist.field.actuator import Actuators
-from artist.field.actuator_ideal import IdealActuators
-from artist.field.actuator_linear import LinearActuators
+from artist.field.actuators import Actuators
+from artist.field.actuators_ideal import IdealActuators
+from artist.field.actuators_linear import LinearActuators
 from artist.field.facets_nurbs import NurbsFacet
 from artist.field.heliostat_field import HeliostatField
 from artist.field.kinematic import Kinematic
