+++ tests/util/test_alignment_optimizer_methods.py
@@ -6,7 +6,7 @@ import torch
 
 from artist import ARTIST_ROOT
 from artist.util import config_dictionary, paint_loader, set_logger_config
-from artist.util.alignment_optimizer import KinematicOptimizer
+from artist.util.kinematic_optimizer import KinematicOptimizer
 from artist.util.scenario import Scenario
 
 # Set up logger.
